webdriver_server

Module actions

Structs

PendingPointerMove 🔒

PendingScroll 🔒

https://github.com/w3c/webdriver/issues/1952 For some unknown reason, unlike Pointer Actions, Wheel Actions in spec has precision of integer. We just use double precision, since our WheelEvent also uses double precision.

PointerInputState 🔒

https://w3c.github.io/webdriver/#dfn-pointer-input-source

Enums

ActionItem 🔒

InputSourceState 🔒

https://w3c.github.io/webdriver/#dfn-input-source-state

PendingActions 🔒

Statics

ELEMENT_CLICK_BUTTON 🔒

https://w3c.github.io/webdriver/#dfn-element-click This is hard-coded as 0 in spec.

MOVESCROLL_INTERVAL 🔒

Interval between wheelScroll and pointerMove increments in ms, based on common vsync

Functions

compute_tick_duration 🔒

touch_pointer_type_for 🔒

https://w3c.github.io/webdriver/#dfn-computing-the-tick-duration Map the webdriver-side PointerType to the embedder-side TouchPointerType. Only valid for Pen and Touch; Mouse should never reach the touch path.

Type Aliases

ActionsByTick 🔒

Consumed by the dispatch_actions method.

TickActions 🔒

A set of actions with multiple sources executed within a single tick. The id is used to identify the source of the actions.