webdriver_server

Struct Handler

pub(crate) struct Handler {
    pub(crate) load_status_receiver: RoutedReceiver<WebDriverLoadStatus>,
    pub(crate) load_status_sender: GenericSender<WebDriverLoadStatus>,
    pub(crate) session: Option<WebDriverSession>,
    pub(crate) embedder_sender: Sender<WebDriverCommandMsg>,
    pub(crate) event_loop_waker: Box<dyn EventLoopWaker>,
    pub(crate) pending_input_event_receivers: Vec<Receiver<()>>,
    pub(crate) pending_actions: Vec<PendingActions>,
    pub(crate) default_preferences: Preferences,
}

Fields
load_status_receiver: RoutedReceiver<WebDriverLoadStatus>

The threaded receiver on which we can block for a load-status. It will receive messages sent on the load_status_sender, and forwarded by the IPC router.

load_status_sender: GenericSender<WebDriverLoadStatus>

The IPC sender which we can clone and pass along to the constellation, for it to send us a load-status. Messages sent on it will be forwarded to the load_status_receiver.

session: Option<WebDriverSession>embedder_sender: Sender<WebDriverCommandMsg>

A Sender that sends messages to the embedder that this `WebDriver instance controls. In addition to sending a message, we must always wake up the embedder’s event loop so it knows that more messages are available for processing.

event_loop_waker: Box<dyn EventLoopWaker>

An EventLoopWaker which is used to wake up the embedder event loop.

pending_input_event_receivers: Vec<Receiver<()>>

A list of Receivers that are used to track when input events are handled in the DOM. Once these receivers receive a response, we know that the event has been handled.

TODO: Once we upgrade crossbeam-channel this can be replaced with a WaitGroup.

pending_actions: Vec<PendingActions>

Ongoing PointerMoveAction or [WheelScrollAction] that are being incrementally processed across multiple execution cycles within the current tick.

default_preferences: Preferences

The base set of preferences to treat as default when resetting.

Implementations

impl Handler

pub(crate) fn dispatch_actions( &mut self, actions_by_tick: Vec<Vec<(String, ActionItem)>>, browsing_context: BrowsingContextId, ) -> Result<(), ErrorStatus>

https://w3c.github.io/webdriver/#dfn-dispatch-actions https://w3c.github.io/webdriver/#dfn-dispatch-actions-inner For Servo, “dispatch actions” is identical to “dispatch actions inner”, as they are only different for a session that can run commands in parallel.

fn process_pending_actions(&mut self, tick_start: &Instant)

https://w3c.github.io/webdriver/#dfn-perform-a-pointer-move Step 9. Run the following substeps in parallel: Step 9.1. Asynchronously wait for an implementation defined amount of time to pass. Step 9.2. Perform a pointer move with arguments input state, duration, start x, start y, target x, target y. https://w3c.github.io/webdriver/#dfn-perform-a-scroll Step 7. Run the following substeps in parallel: Step 7.1. Asynchronously wait for an implementation defined amount of time to pass. Step 7.2. Perform a scroll with arguments duration, x, y, target delta x, target delta y, current delta x, current delta y.

fn wait_for_input_event_handled(&mut self) -> Result<(), ErrorStatus>

fn dispatch_tick_actions( &mut self, tick_actions: &Vec<(String, ActionItem)>, tick_duration: u64, tick_start: &Instant, ) -> Result<(), ErrorStatus>

https://w3c.github.io/webdriver/#dfn-dispatch-tick-actions

fn dispatch_pause_action(&mut self, _input_id: &str)

https://w3c.github.io/webdriver/#dfn-dispatch-a-pause-action This is a dummy action. The sole effect of the action is that its duration parameter may decide “tick duration”.

fn dispatch_keydown_action(&mut self, input_id: &str, action: &KeyDownAction)

https://w3c.github.io/webdriver/#dfn-dispatch-a-keydown-action

fn dispatch_keyup_action(&mut self, input_id: &str, action: &KeyUpAction)

https://w3c.github.io/webdriver/#dfn-dispatch-a-keyup-action

fn dispatch_pointercancel_action(&mut self, source_id: &str)

https://w3c.github.io/webdriver/#dfn-dispatch-a-pointercancel-action

fn dispatch_pointerdown_action( &mut self, input_id: &str, action: &PointerDownAction, )

https://w3c.github.io/webdriver/#dfn-dispatch-a-pointerdown-action

fn dispatch_pointerup_action( &mut self, input_id: &str, action: &PointerUpAction, )

https://w3c.github.io/webdriver/#dfn-dispatch-a-pointerup-action

fn dispatch_pointermove_action( &mut self, input_id: &str, action: &PointerMoveAction, tick_duration: u64, tick_start: &Instant, ) -> Result<(), ErrorStatus>

https://w3c.github.io/webdriver/#dfn-dispatch-a-pointermove-action

fn perform_pointer_move( &mut self, input_id: &str, duration: u64, start_x: f64, start_y: f64, target_x: f64, target_y: f64, tick_start: &Instant, )

https://w3c.github.io/webdriver/#dfn-perform-a-pointer-move

fn dispatch_scroll_action( &mut self, input_id: &str, action: &WheelScrollAction, tick_duration: u64, tick_start: &Instant, ) -> Result<(), ErrorStatus>

https://w3c.github.io/webdriver/#dfn-dispatch-a-scroll-action

fn perform_scroll( &mut self, input_id: &str, duration: u64, x: f64, y: f64, target_delta_x: f64, target_delta_y: f64, current_delta_x: f64, current_delta_y: f64, tick_start: &Instant, )

https://w3c.github.io/webdriver/#dfn-perform-a-scroll

fn check_viewport_bound(&self, x: f64, y: f64) -> Result<(), ErrorStatus>

Verify that the given coordinates are within the boundary of the viewport. If x or y is less than 0 or greater than the width of the viewport in CSS pixels, then return error with error code move target out of bounds.

pub(crate) fn get_origin_relative_coordinates( &self, origin: &PointerOrigin, x_offset: f64, y_offset: f64, input_id: &str, ) -> Result<(f64, f64), ErrorStatus>

https://w3c.github.io/webdriver/#dfn-get-coordinates-relative-to-an-origin

fn get_element_in_view_center_point( &self, web_element: &WebElement, ) -> Result<(i64, i64), ErrorStatus>

https://w3c.github.io/webdriver/#dfn-center-point

pub(crate) fn extract_an_action_sequence( &mut self, actions: Vec<ActionSequence>, ) -> Vec<Vec<(String, ActionItem)>>

https://w3c.github.io/webdriver/#dfn-extract-an-action-sequence

fn process_an_input_source_action_sequence( &mut self, action_sequence: ActionSequence, ) -> Vec<ActionItem>

https://w3c.github.io/webdriver/#dfn-process-an-input-source-action-sequence

fn input_state_table_mut(&mut self) -> &mut HashMap<String, InputSourceState>

fn get_pointer_input_state_mut( &mut self, input_id: &str, ) -> &mut PointerInputState

fn get_pointer_input_state(&self, input_id: &str) -> &PointerInputState

impl Handler

pub(crate) fn extract_script_arguments( &self, parameters: JavascriptCommandParameters, ) -> WebDriverResult<(String, Vec<String>)>

https://w3c.github.io/webdriver/#dfn-extract-the-script-arguments-from-a-request

fn deserialize_web_element(&self, element: &Value) -> WebDriverResult<String>

https://w3c.github.io/webdriver/#dfn-deserialize-a-web-element

fn deserialize_shadow_root( &self, shadow_root: &Value, ) -> WebDriverResult<String>

https://w3c.github.io/webdriver/#dfn-deserialize-a-shadow-root

fn deserialize_web_frame(&self, frame: &Value) -> WebDriverResult<String>

https://w3c.github.io/webdriver/#dfn-deserialize-a-web-frame

fn deserialize_web_window(&self, window: &Value) -> WebDriverResult<String>

https://w3c.github.io/webdriver/#dfn-deserialize-a-web-window

fn json_deserialize(&self, v: &Value) -> WebDriverResult<String>

https://w3c.github.io/webdriver/#dfn-json-deserialize

impl Handler

pub(crate) fn create_session( &mut self, capabilities: &mut Map<String, Value>, servo_capabilities: &ServoCapabilities, ) -> WebDriverResult<Uuid>

https://w3c.github.io/webdriver/#dfn-create-a-session

impl Handler

pub(crate) fn handle_dismiss_alert(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dismiss-alert

pub(crate) fn handle_accept_alert(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#accept-alert

pub(crate) fn handle_get_alert_text(&self) -> WebDriverResult<WebDriverResponse>

https://www.w3.org/TR/webdriver2/#dfn-get-alert-text

pub(crate) fn handle_send_alert_text( &self, text: String, ) -> WebDriverResult<WebDriverResponse>

https://www.w3.org/TR/webdriver2/#dfn-send-alert-text

pub(crate) fn handle_any_user_prompts( &self, webview_id: WebViewId, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-handle-any-user-prompts

impl Handler

pub(crate) fn new( embedder_sender: Sender<WebDriverCommandMsg>, event_loop_waker: Box<dyn EventLoopWaker>, default_preferences: Preferences, ) -> Handler

pub(crate) fn browsing_context_id(&self) -> WebDriverResult<BrowsingContextId>

pub(crate) fn webview_id(&self) -> WebDriverResult<WebViewId>

pub(crate) fn send_input_event_to_embedder(&self, input_event: InputEvent)

pub(crate) fn send_blocking_input_event_to_embedder( &mut self, input_event: InputEvent, )

pub(crate) fn send_message_to_embedder( &self, msg: WebDriverCommandMsg, ) -> WebDriverResult<()>

pub(crate) fn add_load_status_sender(&self) -> WebDriverResult<()>

pub(crate) fn clear_load_status_sender(&self) -> WebDriverResult<()>

pub(crate) fn verified_webview_id(&self) -> WebViewId

pub(crate) fn focused_webview_id(&self) -> WebDriverResult<Option<WebViewId>>

pub(crate) fn session(&self) -> WebDriverResult<&WebDriverSession>

pub(crate) fn session_mut(&mut self) -> WebDriverResult<&mut WebDriverSession>

pub(crate) fn handle_new_session( &mut self, parameters: &NewSessionParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#new-session

pub(crate) fn handle_delete_session( &mut self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-delete-session

pub(crate) fn handle_status(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#status

pub(crate) fn browsing_context_script_command( &self, cmd_msg: WebDriverScriptCommand, verify: VerifyBrowsingContextIsOpen, ) -> WebDriverResult<()>

Send command to Script Thread with session’s current browsing context. If verify is VerifyBrowsingContextIsOpen::Yes, it would verify the existence of browsing context before sending.

pub(crate) fn top_level_script_command( &self, cmd_msg: WebDriverScriptCommand, verify: VerifyBrowsingContextIsOpen, ) -> WebDriverResult<()>

Send command to Script Thread with session’s current top-level browsing context. If verify is VerifyBrowsingContextIsOpen::Yes, it would verify the existence of top-level browsing context before sending.

pub(crate) fn handle_get( &mut self, parameters: &GetParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#navigate-to

pub(crate) fn wait_document_ready( &self, timeout: Option<u64>, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn wait_for_navigation_complete( &self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-wait-for-navigation-to-complete

pub(crate) fn wait_for_navigation(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-wait-for-navigation-to-complete

pub(crate) fn handle_current_url(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-get-current-url

pub(crate) fn handle_window_rect( &self, verify: VerifyBrowsingContextIsOpen, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-window-rect

pub(crate) fn handle_set_window_rect( &self, params: &WindowRectParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#set-window-rect

pub(crate) fn handle_maximize_window( &mut self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#maximize-window

pub(crate) fn handle_is_enabled( &self, element: &WebElement, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_is_selected( &self, element: &WebElement, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_go_back(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#back

pub(crate) fn handle_go_forward(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#forward

pub(crate) fn handle_refresh(&mut self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#refresh

pub(crate) fn handle_title(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-title

pub(crate) fn handle_window_handle( &mut self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-window-handle

pub(crate) fn handle_window_handles( &mut self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-window-handles

pub(crate) fn get_window_handle( &mut self, webview_id: WebViewId, ) -> Option<String>

pub(crate) fn get_window_handles(&self) -> Vec<String>

pub(crate) fn get_all_webview_ids(&self) -> Vec<WebViewId>

pub(crate) fn handle_close_window( &mut self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#close-window

pub(crate) fn handle_new_window( &mut self, parameters: &NewWindowParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#new-window

pub(crate) fn handle_switch_to_frame( &mut self, parameters: &SwitchToFrameParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-switch-to-frame

pub(crate) fn handle_switch_to_parent_frame( &mut self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#switch-to-parent-frame

pub(crate) fn handle_switch_to_window( &mut self, parameters: &SwitchToWindowParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#switch-to-window

pub(crate) fn switch_to_frame( &mut self, frame_id: WebDriverFrameId, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_find_element( &self, parameters: &LocatorParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#find-element

pub(crate) fn implicit_wait<T>( &self, callback: impl Fn() -> Result<(bool, T), (bool, WebDriverError)>, ) -> Result<T, WebDriverError>

The boolean in callback result indicates whether implicit_wait can early return before timeout with current result.

pub(crate) fn handle_find_elements( &self, parameters: &LocatorParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#find-elements

pub(crate) fn handle_find_element_from_element( &self, element: &WebElement, parameters: &LocatorParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#find-element-from-element

pub(crate) fn handle_find_elements_from_element( &self, element: &WebElement, parameters: &LocatorParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#find-elements-from-element

pub(crate) fn handle_find_elements_from_shadow_root( &self, shadow_root: &ShadowRoot, parameters: &LocatorParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#find-elements-from-shadow-root

pub(crate) fn handle_find_element_from_shadow_root( &self, shadow_root: &ShadowRoot, parameters: &LocatorParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#find-element-from-shadow-root

pub(crate) fn handle_get_shadow_root( &self, element: WebElement, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-element-shadow-root

pub(crate) fn handle_element_rect( &self, element: &WebElement, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-element-rect

pub(crate) fn handle_element_text( &self, element: &WebElement, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-get-element-text

pub(crate) fn handle_active_element(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-active-element

pub(crate) fn handle_computed_role( &self, element: &WebElement, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-computed-role

pub(crate) fn handle_element_tag_name( &self, element: &WebElement, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-element-tag-name

pub(crate) fn handle_element_attribute( &self, element: &WebElement, name: &str, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-element-attribute

pub(crate) fn handle_element_property( &self, element: &WebElement, name: &str, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-element-property

pub(crate) fn handle_element_css( &self, element: &WebElement, name: &str, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-element-css-value

pub(crate) fn handle_get_cookies(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-all-cookies

pub(crate) fn handle_get_cookie( &self, name: String, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-named-cookie

pub(crate) fn handle_add_cookie( &self, params: &AddCookieParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#add-cookie

pub(crate) fn handle_delete_cookie( &self, name: String, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#delete-cookie

pub(crate) fn handle_delete_cookies(&self) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#delete-all-cookies

pub(crate) fn handle_get_timeouts( &mut self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#get-timeouts

pub(crate) fn handle_set_timeouts( &mut self, parameters: &TimeoutsParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#set-timeouts

pub(crate) fn handle_get_page_source( &self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-get-page-source

pub(crate) fn handle_perform_actions( &mut self, parameters: ActionsParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#perform-actions

pub(crate) fn handle_release_actions( &mut self, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-release-actions

pub(crate) fn handle_execute_script( &self, parameters: JavascriptCommandParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-execute-script

pub(crate) fn handle_execute_async_script( &self, parameters: JavascriptCommandParameters, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn javascript_evaluation_result_to_webdriver_response( &self, result: WebDriverJSResult, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_element_send_keys( &mut self, element: &WebElement, keys: &SendKeysParameters, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#dfn-element-send-keys

pub(crate) fn handle_element_clear( &self, element: &WebElement, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#element-clear

pub(crate) fn handle_element_click( &mut self, element: &WebElement, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#element-click

pub(crate) fn perform_element_click( &mut self, element: String, ) -> WebDriverResult<WebDriverResponse>

https://w3c.github.io/webdriver/#element-click Step 8 for elements other than ,

pub(crate) fn take_screenshot( &self, rect: Option<Rect<f32, CSSPixel>>, ) -> WebDriverResult<String>

pub(crate) fn handle_take_screenshot( &self, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_take_element_screenshot( &self, element: &WebElement, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_custom_handlers_set_mode( &self, parameters: &CustomHandlersSetModeParameters, ) -> WebDriverResult<WebDriverResponse>

https://html.spec.whatwg.org/multipage/#set-rph-registration-mode

pub(crate) fn handle_get_prefs( &self, parameters: &GetPrefsParameters, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_set_prefs( &self, parameters: &SetPrefsParameters, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_reset_prefs( &self, parameters: &GetPrefsParameters, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_shutdown(&self) -> WebDriverResult<WebDriverResponse>

pub(crate) fn handle_reset_all_cookies( &self, ) -> WebDriverResult<WebDriverResponse>

pub(crate) fn verify_top_level_browsing_context_is_open( &self, webview_id: WebViewId, ) -> Result<(), WebDriverError>

pub(crate) fn verify_browsing_context_is_open( &self, browsing_context_id: BrowsingContextId, ) -> Result<(), WebDriverError>

pub(crate) fn wait_until_browsing_context_is_open( &self, browsing_context_id: BrowsingContextId, ) -> WebDriverResult<()>

pub(crate) fn focus_webview(&self, webview_id: WebViewId) -> WebDriverResult<()>

pub(crate) fn focus_browsing_context( &self, browsing_cotext_id: BrowsingContextId, ) -> WebDriverResult<()>