Struct webdriver_server::WebDriverSession

struct WebDriverSession {
    id: Uuid,
    frame_id: Option<FrameId>,
    script_timeout: Option<u32>,
    load_timeout: u32,
    implicit_wait_timeout: u32,
}

Represents the current WebDriver session and holds relevant session state.

Fields
id: Uuid frame_id: Option<FrameId> script_timeout: Option<u32>

Time to wait for injected scripts to run before interrupting them. A [None] value specifies that the script should run indefinitely.

load_timeout: u32

Time to wait for a page to finish loading upon navigation.

implicit_wait_timeout: u32

Time to wait for the element location strategy when retrieving elements, and when waiting for an element to become interactable.

Methods

impl WebDriverSession

fn new() -> WebDriverSession