webdriver_server

Module timeout

Structs

TimeoutsConfiguration 🔒

https://w3c.github.io/webdriver/#dfn-timeouts-configuration A None timeout means waiting indefinitely.

Constants

DEFAULT_IMPLICIT_WAIT 🔒

Initial initial wait timeout from https://w3c.github.io/webdriver/#dfn-timeouts-configuration.

DEFAULT_PAGE_LOAD_TIMEOUT 🔒

Initial page load timeout from https://w3c.github.io/webdriver/#dfn-timeouts-configuration.

DEFAULT_SCRIPT_TIMEOUT 🔒

Initial script timeout from https://w3c.github.io/webdriver/#dfn-timeouts-configuration.

SCREENSHOT_TIMEOUT 🔒

An amount of time to wait before considering that a screenshot has timed out. If after 10 seconds the screenshot cannot be taken, assume that the test has timed out.

Functions

deserialize_as_timeouts_configuration 🔒

https://w3c.github.io/webdriver/#dfn-deserialize-as-timeouts-configuration

serialize_timeouts_configuration 🔒

https://w3c.github.io/webdriver/#dfn-serialize-the-timeouts-configuration