webdriver_server

Module actions

Structs

PointerInputState 🔒

https://w3c.github.io/webdriver/#dfn-pointer-input-source

Enums

ActionItem 🔒

InputSourceState 🔒

https://w3c.github.io/webdriver/#dfn-input-source-state

Statics

ELEMENT_CLICK_BUTTON 🔒

https://w3c.github.io/webdriver/#dfn-element-click This is hard-coded as 0 in spec.

MAXIMUM_SAFE_INTEGER 🔒

https://262.ecma-international.org/6.0/#sec-number.max_safe_integer 2^53 - 1

POINTERMOVE_INTERVAL 🔒

Interval between wheelScroll and pointerMove increments in ms, based on common vsync

WHEELSCROLL_INTERVAL 🔒

Functions

compute_tick_duration 🔒

https://w3c.github.io/webdriver/#dfn-computing-the-tick-duration

Type Aliases

ActionsByTick 🔒

Consumed by the dispatch_actions method.

TickActions 🔒

A set of actions with multiple sources executed within a single tick. The id is used to identify the source of the actions.